refactor Calculator and Water Bottle tasks

+++ src/main/java/Calculator.java
@@ -1,29 +1,29 @@
 public class Calculator {
-    private int number1;
-    private int number2;
+    private double number1;
+    private double number2;
 
-    public Calculator(int number1, int number2){
+    public Calculator(double number1, double number2){
         this.number1 = number1;
         this.number2 = number2;
     }
 
-    public int add(){
-        int sum = this.number1 + this.number2;
+    public double add(){
+        double sum = this.number1 + this.number2;
         return sum;
     }
 
-    public int subtract(){
-        int result = this.number1 - this.number2;
+    public double subtract(){
+        double result = this.number1 - this.number2;
         return result;
     }
 
-    public int multiply(){
-        int result = this.number1 * this.number2;
+    public double multiply(){
+        double result = this.number1 * this.number2;
         return result;
     }
 
-    public int divide(){
-        int result = this.number1 / this.number2;
+    public double divide(){
+        double result = this.number1 / this.number2;
         return result;
     }
